To_do_project/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
django-filter==2.4.0
django-floppyforms==1.9.0
fpdf==1.7.2
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
